fix: 예약 날짜 검증을 서비스로 위임
