+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -2,13 +2,9 @@ package roomescape.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import roomescape.exception.BadRequestException;
-
-import java.time.LocalDate;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static roomescape.TestFixture.*;
@@ -23,21 +19,5 @@ class ReservationTest {
         // when & then
         assertThatThrownBy(() -> new Reservation(invalidName, MIA_RESERVATION_DATE, new ReservationTime(MIA_RESERVATION_TIME), WOOTECO_THEME()))
                 .isInstanceOf(BadRequestException.class);
-    }
-
-    @ParameterizedTest
-    @MethodSource("invalidLocalDate")
-    @DisplayName("예약 날짜는 현재 날짜 이후이다.")
-    void validateDate(LocalDate invalidDate) {
-        // when & then
-        assertThatThrownBy(() -> new Reservation(USER_MIA, invalidDate, new ReservationTime(MIA_RESERVATION_TIME), WOOTECO_THEME()))
-                .isInstanceOf(BadRequestException.class);
-    }
-
-    private static Stream<LocalDate> invalidLocalDate() {
-        return Stream.of(
-                LocalDate.now(),
-                LocalDate.now().minusDays(1L)
-        );
     }
 }
